requirements.txt
# Function dependencies, for example:
# package>=version
# smartsheet-python-sdk>=2.0
# mysql-connector-python==8.0.11
google-cloud-pubsub==1.1.0
